prevented services from being overwritten

+++ Scripts/make_services_4.sh
@@ -3,7 +3,12 @@
 #these services will run the vncservers for each VNC user at boot
 while IFS=':' read -r index email || [[ -n "$email" ]]; do
 	USER=${email%@*}
-	sed -e "s/TMP/$USER/g" ../Config/uservnc@.service >  ../Config/${USER}vnc@.service
-	chmod 755 ../Config/${USER}vnc@.service
-	cp ../Config/${USER}vnc@.service /etc/systemd/system/
+	if [ -f ../Config/${USER}vnc@.service ]
+	then
+		echo "../Config/${USER}vnc@.service already exists..."
+	else
+		sed -e "s/TMP/$USER/g" ../Config/uservnc@.service >  ../Config/${USER}vnc@.service
+		chmod 755 ../Config/${USER}vnc@.service
+		cp ../Config/${USER}vnc@.service /etc/systemd/system/
+	fi
 done < ../Config/VNC_users.txt
